abi # Avoid gcc/clang ABI incompatibility caused by PlacementNew
Avoid gcc/clang ABI incompatibility caused by PlacementNew

GCC and clang do not use same calling convention for passing empty struct
parameters. There is more information about this in
https://itanium-cxx-abi.github.io/cxx-abi/cxx-abi-dev/archives/2015-December/002869.html

Unfortunately this can create an issue in capnproto if it is built without
optimizations in GCC and the resulting static libraries are used in a clang
program, or vice versa.

Depending on what order libraries are specified on the linker command line and
whether the code compiled with the other compiler is calling any header
functions that cause weak a `operator new(unsigned int, kj::_::PlacementNew,
void*)` symbol being defined in its own objects, this can cause the linker to
make a clang generated `kj::ctor` function call a GCC generated `operator new`
function and the resulting program to crash due to the compilers using
different calling conventions for `operator new`.

This problem is difficult to avoid in general, but pretty easy to avoid here by
changing `operator new` parameter order so the empty struct parameter is last.

This change should be beneficial for capnproto users that may be compiling it
without optimizations and not necessarily using a single compiler to build all
their dependencies.

The problem does not occur if any optimizations are enabled because `operator
new` calls are inlined in that case.